requirements.txt
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.1.3
fonttools==4.45.0
h11==0.14.0
idna==3.4
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
zipp==3.17.0
